rename stats_scalar
