+++ source/consumers/StatsScalar.h
